requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
numpy==1.19.2
pandas==1.1.3
prompt-toolkit==1.0.14
Pygments==2.7.1
PyInquirer==1.0.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
regex==2020.10.23
requests==2.24.0
six==1.15.0
urllib3==1.25.11
wcwidth==0.2.5
